language/demo/streamlit/document-qa/requirements.txt
google-cloud-aiplatform
google-cloud-documentai
backoff
ruamel.yaml
streamlit
streamlit-chat
pandas
PyPDF2
chromadb
sequence-evaluate
sentence-transformers 
rouge